rename module training_corpus to corpus
