+++ test/test_training.py
@@ -9,7 +9,7 @@ from xnmt.lstm import BiLSTMSeqTransducer
 from xnmt.pyramidal import PyramidalLSTMSeqTransducer
 from xnmt.attender import MlpAttender, DotAttender
 from xnmt.decoder import MlpSoftmaxDecoder, CopyBridge
-from xnmt.training_corpus import BilingualTrainingCorpus
+from xnmt.corpus import BilingualTrainingCorpus
 from xnmt.input import BilingualCorpusParser, PlainTextReader
 from xnmt.batcher import mark_as_batch, Mask, SrcBatcher
 import xnmt.train
